requirements.txt
fastapi
uvicorn
websockets
yt-dlp
opencv-python-headless
numpy
mediapipe
ffmpeg-python
jinja2
python-multipart
